requirements.txt
bcrypt==4.0.1
coverage==6.5.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
marshmallow==3.18.0
nose==1.3.7
PyJWT==2.6.0
pyodbc==4.0.35
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
psycopg2==2.9.1